Change csv import to take show details not show_id
Add a method for Show which searches for a show by title and either returns it
if found or creates a new show to be returned from the parameters passed in.
Use this method in the performance import model to allow the CSV importing of
performances which have not had a show created yet.